backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.10.0
python-multipart==0.0.6
asyncpg==0.29.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
aiofiles==24.1.0
numpy==1.24.3
anthropic==0.37.1
PyYAML==6.0.1
facebook-business==19.0.3
cryptography==41.0.7
apscheduler==3.10.4
boto3==1.34.0
